requirements.txt
git+https://github.com/aaroswings/albumentations.git@random-tone-curve-quick-fix
numpy==1.19.4
opencv-python==4.2.0.32
opencv-python-headless==4.5.1.48
pandas==1.1.5
Pillow==8.0.1
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.4
seaborn==0.11.1
torch==1.7.1
torchvision==0.8.2
tqdm==4.55.0
